Get commit history from bacchus1
The original ibe commit history was automatically created from accurev,
and it had the strange by-product of creating 60+ concurrent branches
to add in each commit. This history is only confusing, and not at all
helpful, so it had been banished to
bacchus1:/stage/irsa-vcs/git/ibe.git

If you truly wish to toil through that history, then check out this
"history" branch from that repo.